Fix typo in protobuf. Add expected_keepers and replace wasm permissioned keeper with our interface. Move ibc_middleware to prevent circular dependency.
