+++ x/ibcratelimit/expected_keepers.go
@@ -1,0 +1,7 @@
+package ibcratelimit
+
+import sdk "github.com/cosmos/cosmos-sdk/types"
+
+type PermissionedKeeper interface {
+	Sudo(ctx sdk.Context, contractAddress sdk.AccAddress, msg []byte) ([]byte, error)
+}
