+++ x/ibcratelimit/module/ibc_middleware_test.go
@@ -1,4 +1,4 @@
-package ibcratelimit_test
+package module_test
 
 import (
 	"encoding/json"
@@ -46,6 +46,10 @@ type MiddlewareTestSuite struct {
 
 // TODO Check this
 //var oldConsensusMinFee = txfeetypes.ConsensusMinFee
+
+const (
+	ByteCodePath = "../bytecode/rate_limiter.wasm"
+)
 
 // Setup
 func TestMiddlewareTestSuite(t *testing.T) {
@@ -353,7 +357,7 @@ func (suite *MiddlewareTestSuite) fullSendTest(native bool) map[string]string {
 	fmt.Printf("Testing send rate limiting for denom=%s, channelValue=%s, quota=%s, sendAmount=%s\n", denom, channelValue, quota, sendAmount)
 
 	// Setup contract
-	suite.chainA.StoreContractCodeDirect(&suite.Suite, "./bytecode/rate_limiter.wasm")
+	suite.chainA.StoreContractCodeDirect(&suite.Suite, ByteCodePath)
 	quotas := suite.BuildChannelQuota("weekly", channel, denom, 604800, 5, 5)
 	fmt.Println(quotas)
 	addr := suite.chainA.InstantiateRLContract(&suite.Suite, quotas)
@@ -448,7 +452,7 @@ func (suite *MiddlewareTestSuite) fullRecvTest(native bool) {
 	fmt.Printf("Testing recv rate limiting for denom=%s, channelValue=%s, quota=%s, sendAmount=%s\n", localDenom, channelValue, quota, sendAmount)
 
 	// Setup contract
-	suite.chainA.StoreContractCodeDirect(&suite.Suite, "./bytecode/rate_limiter.wasm")
+	suite.chainA.StoreContractCodeDirect(&suite.Suite, ByteCodePath)
 	quotas := suite.BuildChannelQuota("weekly", channel, localDenom, 604800, 4, 4)
 	addr := suite.chainA.InstantiateRLContract(&suite.Suite, quotas)
 	suite.chainA.RegisterRateLimitingContract(addr)
@@ -484,7 +488,7 @@ func (suite *MiddlewareTestSuite) TestRecvTransferWithRateLimitingNonNative() {
 // Test no rate limiting occurs when the contract is set, but not quotas are condifured for the path
 func (suite *MiddlewareTestSuite) TestSendTransferNoQuota() {
 	// Setup contract
-	suite.chainA.StoreContractCodeDirect(&suite.Suite, "./bytecode/rate_limiter.wasm")
+	suite.chainA.StoreContractCodeDirect(&suite.Suite, ByteCodePath)
 	addr := suite.chainA.InstantiateRLContract(&suite.Suite, ``)
 	suite.chainA.RegisterRateLimitingContract(addr)
 
@@ -498,7 +502,7 @@ func (suite *MiddlewareTestSuite) TestSendTransferNoQuota() {
 func (suite *MiddlewareTestSuite) TestFailedSendTransfer() {
 	suite.initializeEscrow()
 	// Setup contract
-	suite.chainA.StoreContractCodeDirect(&suite.Suite, "./bytecode/rate_limiter.wasm")
+	suite.chainA.StoreContractCodeDirect(&suite.Suite, ByteCodePath)
 	quotas := suite.BuildChannelQuota("weekly", "channel-0", sdk.DefaultBondDenom, 604800, 1, 1)
 	addr := suite.chainA.InstantiateRLContract(&suite.Suite, quotas)
 	suite.chainA.RegisterRateLimitingContract(addr)
@@ -568,7 +572,7 @@ func (suite *MiddlewareTestSuite) TestFailedSendTransfer() {
 
 func (suite *MiddlewareTestSuite) TestUnsetRateLimitingContract() {
 	// Setup contract
-	suite.chainA.StoreContractCodeDirect(&suite.Suite, "./bytecode/rate_limiter.wasm")
+	suite.chainA.StoreContractCodeDirect(&suite.Suite, ByteCodePath)
 	addr := suite.chainA.InstantiateRLContract(&suite.Suite, "")
 	suite.chainA.RegisterRateLimitingContract(addr)
